trnsvr/requirements.txt
flask==1.1.2
psycopg2==2.7.6.1
pytest==3.0.5
pytest-cov==2.4.0
tensorflowjs==3.4.0